requirements.txt
fastapi
uvicorn
python-dotenv
langchain==0.3.0
langchain-openai
langchain-community
pypdf
tiktoken
faiss-cpu